# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Pollar API",
    "description": "REST API for a voting application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Auth",
      "description": "User authentication"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint para cadastrar um usuário.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Informações do usuário.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Falha na validação"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint para autenticar um usuário.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Informações de login do usuário.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "type": "string",
                  "example": "username/email"
                },
                "password": {
                  "type": "string",
                  "example": "secret"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário autenticado.",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                },
                "userID": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/users/edit/{userId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint para editar um usuário.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário."
          },
          {
            "name": "user",
            "in": "body",
            "description": "Informações do usuário.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string",
                  "example": "Doe"
                },
                "email": {
                  "type": "string",
                  "example": "johndoe@email.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário editado.",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/profile/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint para buscar um usuário.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário buscado.",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/polls/all": {
      "get": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para mostrar todas votações de um usuário.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Votações buscadas.",
            "schema": {
              "$ref": "#/definitions/poll"
            }
          },
          "404": {
            "description": "Votação não encontrada."
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/polls/getPoll/{pollId}": {
      "get": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para buscar uma votações.",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da votação."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/poll"
            },
            "description": "Votação buscada."
          },
          "404": {
            "description": "Votação não encontrada."
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/polls/create": {
      "post": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para criar uma votação.",
        "parameters": [
          {
            "name": "poll",
            "in": "body",
            "description": "Informações da votação.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/poll"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/polls/edit/{pollId}": {
      "put": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para editar uma votação.",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "poll",
            "in": "body",
            "description": "Informações da votação.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/poll"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Votação não encontrada."
          },
          "422": {
            "description": "Não é possível editar uma votação iniciada."
          }
        }
      }
    },
    "/polls/delete/{pollId}": {
      "delete": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para deletar uma votação.",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da votação."
          }
        ],
        "responses": {
          "403": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Votação não encontrada."
          }
        }
      }
    },
    "/polls/start/{pollId}": {
      "patch": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para abrir uma votação para votos.",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da votação."
          }
        ],
        "responses": {
          "403": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Votação não encontrada."
          },
          "422": {
            "description": "Não é possível editar uma votação iniciada."
          }
        }
      }
    },
    "/polls/finish/{pollId}": {
      "patch": {
        "tags": [
          "Polls"
        ],
        "description": "Endpoint para fechar uma votação para votos.",
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da votação."
          }
        ],
        "responses": {
          "403": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Votação não encontrada."
          },
          "422": {
            "description": "Não é possível editar uma votação iniciada."
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "username": {
          "type": "string",
          "example": "jhonDoe"
        },
        "first_name": {
          "type": "string",
          "example": "Jhon"
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "jhon.doe@email.com"
        },
        "password": {
          "type": "string",
          "example": "secret"
        },
        "cpf": {
          "type": "string",
          "example": "12345678901"
        },
        "activated": {
          "type": "number",
          "example": 1
        },
        "createdAt": {
          "type": "string",
          "example": "2022-01-01 00:00:00"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-01-01 00:00:00"
        }
      }
    },
    "Poll": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "teste"
        },
        "initial_date": {
          "type": "string",
          "example": "2022-01-01 00:00:00"
        },
        "end_date": {
          "type": "string",
          "example": "2022-01-02 00:00:00"
        },
        "sharing_url": {
          "type": "string",
          "example": "/s9Xjk23Ac"
        },
        "description": {
          "type": "string",
          "example": "Descrição"
        },
        "started": {
          "type": "number",
          "example": 0
        },
        "finished": {
          "type": "number",
          "example": 0
        },
        "activated": {
          "type": "number",
          "example": 1
        },
        "createdAt": {
          "type": "string",
          "example": "2022-01-01 00:00:00"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-01-01 00:00:00"
        },
        "userId": {
          "type": "number",
          "example": 1
        }
      }
    }
  }
}